install and configure syslog-ng
The configuration causes syslog-ng to listen to port 514/UDP and 601/TCP for syslog messages.
For TCP clients, it expects the 'octet counting' method to delimit messages. This allows multi-line messages to work.

The syslog messages are logged to the /var/log/syslog directory, to a host/day-specific file.

Internal (syslog-ng) messages are logged to /var/log/syslog-ng.log.