add cron-based removal of old logs in /var/log/syslog/
We don't need to do rotation, because syslog-ng itself will start new files every day because we put a timestamp in the filename.

There may be other log files that need to be rotated and/or pruned, but these are going to be the ones that chew up the most disk space by far.