make test script exit with non-zero if test seems to have failed
This is pretty rough, as it assumes that "count=15" in the output means the message were all sent successfully. This will probably be the case.

This doesn't check that syslog-ng wrote the messages to a file, or that splunkd successfully picked it up from there. These must be checked manually, as described in the readme.

This commit also tries to making debugging a little easier, by printing out the stdout/stderr of the loggen command, if there was a failure.