capture syslog-ng debug output
For some reason, even though I've not specified --stderr, syslog-ng is spitting all it debug output to stderr instead of the internal() log source. (The docs seem to say this shouldn't be the case, but I've not really dug into it further).
So instead of relying on internal() (and the associated syslog-ng.log file), we'll just capture stderr via supervisord.

While at it, this also captures splunk and cron's stdout and stderr.
Also, this commit changes the splunk config to track all of these files.

This commit also removes the --no-caps flag because this is the default. I'm not sure why I added it in the beginning.